Remove the other triangle
